Changed to standart_response

+++ IssueTracker/IssueTrackerApp/views.py
@@ -1,6 +1,7 @@
 from django.http.response import Http404
 from .models import Comment, Project, User, Issue, UserProject, Assignee
 from .serializers import IssueSerializer, UserSerializer, ProjectSerializer, CommentSerializer
+from .utils import standard_response
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status, permissions
@@ -12,14 +13,18 @@ class UserList(APIView):
     def get(self, request, format=None):
         users = User.objects.all()
         serializer = UserSerializer(users, many=True, context = {'request':request})
-        return Response(serializer.data)
+        res = standard_response(data=serializer.data)
+        return Response(res)
     
     def post(self, request, format=None):
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status.HTTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res, status=status.HTTP_201_CREATED)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status.HTTTP_400_BAD_REQUEST)
 
 class UserDetail(APIView):
     """
@@ -29,25 +34,45 @@ class UserDetail(APIView):
         try:
             return User.objects.get(pk=pk)
         except User.DoesNotExist:
-            raise Http404
+            return None
 
     def get(self, request, pk, format=None):
         user = self.get_object(pk)
+        if not user:
+            res = standard_response(
+                errors={'error':'The user does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
         serializer = UserSerializer(user, context = {'request':request})
-        return Response(serializer.data)
+        res = standard_response(data=serializer.data)
+        return Response(res)
 
     def put(self, request, pk, format=None):
         user = self.get_object(pk)
+        if not user:
+            res = standard_response(
+                errors={'error':'The user does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
         serializer = UserSerializer(user, data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
         user = self.get_object(pk)
+        if not user:
+            res = standard_response(
+                errors={'error':'The user does not exist'}
+            )
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
         user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(standard_response(), status=status.HTTP_204_NO_CONTENT)
 
 class ProjectList(APIView):
     """
@@ -57,14 +82,18 @@ class ProjectList(APIView):
         user = request.user
         projects = user.projects.all()
         serializer = ProjectSerializer(projects, many=True, context = {'request':request})
-        return Response(serializer.data)
+        res = standard_response(data=serializer.data)
+        return Response(res)
 
     def post(self, request, format=None):
         serializer = ProjectSerializer(data=request.data, context = {'request':request})
         if serializer.is_valid():
             UserProject.objects.create(user=request.user, project = serializer.save())
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res, status=status.HTTP_201_CREATED)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
 
 class ProjectDetail(APIView):
     """
@@ -74,35 +103,51 @@ class ProjectDetail(APIView):
         try:
             return Project.objects.get(pk=pk)
         except Project.DoesNotExist:
-            raise Http404
+            return None
 
     def get(self, request, pk, format=None):
-        user = request.user
         project = self.get_object(pk)
+        if not project:
+            res = standard_response(
+                errors={'error':'The project does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
         serializer = ProjectSerializer(project, context = {'request':request})
-        can_view_project = user.projects.filter(id=project.id).count() > 0
-        if can_view_project:
-            return Response(serializer.data)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
-
+        res = standard_response(data=serializer.data)
+        return Response(res)
+    
     def put(self, request, pk, format=None):
-        user = request.user
         project = self.get_object(pk)
-        serializer = ProjectSerializer(project, data=request.data)
+        if not project:
+            res = standard_response(
+                errors={'error':'The project does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        serializer = ProjectSerializer(project, data=request.data, context={'request':request})
+        user = request.user
         can_modify_project = user.projects.filter(id=project.id).count() > 0
         if serializer.is_valid() and can_modify_project:
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
-        user = request.user
         project = self.get_object(pk)
+        if not project:
+            res = standard_response(
+                errors={'error':'The project does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        user = request.user
         can_delete_project = user.projects.filter(id=project.id).count() > 0
         if can_delete_project:
             project.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response(standard_response(), status=status.HTTP_204_NO_CONTENT)
+        
+        return Response(standard_response(), status=status.HTTP_400_BAD_REQUEST)
 
 class IssueList(APIView):
     """
@@ -112,14 +157,18 @@ class IssueList(APIView):
         user = request.user
         issues = user.issues.all()
         serializer = IssueSerializer(issues, many=True)
-        return Response(serializer.data)
+        res = standard_response(data=serializer.data)
+        return Response(res)
 
     def post(self, request, format=None):
         serializer = IssueSerializer(data=request.data, context={'request':request})
         if serializer.is_valid():
             Assignee.objects.create(user= request.user, issue=serializer.save())
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res, status=status.HTTP_201_CREATED)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
     
 class IssueDetail(APIView):
     """
@@ -129,35 +178,56 @@ class IssueDetail(APIView):
         try:
             return Issue.objects.get(pk=pk)
         except Issue.DoesNotExist:
-            raise Http404
+            return None
 
     def get(self, request, pk, format=None):
-        user = request.user
         issue = self.get_object(pk)
+        if not issue:
+            res = standard_response(
+                errors={'error':'The issue does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        user = request.user
         serializer = IssueSerializer(issue)
         can_view_issue = user.issues.filter(id=issue.id).count() > 0
         if can_view_issue:
-            return Response(serializer.data)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        return Response(standard_response(), status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk, format=None):
-        user = request.user
         issue = self.get_object(pk)
+        if not issue:
+            res = standard_response(
+                errors={'error':'The issue does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+
+        user = request.user
         serializer = IssueSerializer(issue, data=request.data)
         can_modify_issue = user.issues.filter(id=issue.id).count() > 0
         if serializer.is_valid() and can_modify_issue:
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
-        user = request.user
         issue = self.get_object(pk)
+        if not issue:
+            res = standard_response(
+                errors={'error':'The issue does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+            
+        user = request.user
         can_delete_issue = user.issues.filter(id=issue.id).count() > 0
         if can_delete_issue:
             issue.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response(standard_response(), status=status.HTTP_204_NO_CONTENT)
+        
+        return Response(standard_response(), status=status.HTTP_400_BAD_REQUEST)
 
 class CommentList(APIView):
     """
@@ -167,7 +237,8 @@ class CommentList(APIView):
         user = request.user
         comments = user.comments.all()
         serializer = CommentSerializer(comments, many=True, context = {'request':request})
-        return Response(serializer.data)
+        res = standard_response(data=serializer.data)
+        return Response(res)
 
     def post(self, request, format=None):
         serializer = CommentSerializer(data=request.data)
@@ -175,8 +246,11 @@ class CommentList(APIView):
         can_view_issue = request.user.projects.filter(id=issue.project.id).count() > 0
         if serializer.is_valid() and can_view_issue:
             serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res, status=status.HTTP_201_CREATED)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
     
 class CommentDetail(APIView):
     """
@@ -186,32 +260,52 @@ class CommentDetail(APIView):
         try:
             return Comment.objects.get(pk=pk)
         except Comment.DoesNotExist:
-            raise Http404
+            return None
 
     def get(self, request, pk, format=None):
-        user = request.user
         comment = self.get_object(pk)
+        if not comment:
+            res = standard_response(
+                errors={'error':'The comment does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        user = request.user
         serializer = CommentSerializer(comment)
         can_view_comment = user.comments.filter(id=comment.id).count() > 0
         if can_view_comment:
-            return Response(serializer.data)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        return Response(standard_response(), status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk, format=None):
-        user = request.user
         comment = self.get_object(pk)
+        if not comment:
+            res = standard_response(
+                errors={'error':'The comment does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        user = request.user
         serializer = CommentSerializer(comment, data=request.data)
         can_modify_comment = user.comments.filter(id=comment.id).count() > 0
         if serializer.is_valid() and can_modify_comment:
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            res = standard_response(data=serializer.data)
+            return Response(res)
+        
+        res = standard_response(errors=serializer.errors)
+        return Response(res, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
-        user = request.user
         comment = self.get_object(pk)
+        if not comment:
+            res = standard_response(
+                errors={'error':'The comment does not exist'})
+            return Response(res, status=status.HTTP_404_NOT_FOUND)
+        
+        user = request.user
         can_delete_comment = user.comments.filter(id=comment.id).count() > 0
         if can_delete_comment:
             comment.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response(standard_response(), status=status.HTTP_204_NO_CONTENT)
+        return Response(standard_response(), status=status.HTTP_400_BAD_REQUEST)
